+++ cmd/token-fuzzer/main.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"strconv"
 
-	. "github.com/skycoin/cx/cxparsergenerator/cxparsingcompletor"
+	. "github.com/skycoin/cx/cxparser/cxparsingcompletor"
 )
 
 var tbits int = 4
