+++ cx/ast/tostring.go
@@ -3,11 +3,12 @@ package ast
 import (
 	"bytes"
 	"fmt"
-	constants2 "github.com/skycoin/cx/cxparsergenerator/constants"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
+
+	constants2 "github.com/skycoin/cx/cxparser/constants"
 
 	"github.com/skycoin/cx/cx/constants"
 	"github.com/skycoin/cx/cx/helper"
