+++ cx/astapi/expressions_test.go
@@ -6,7 +6,7 @@ import (
 	cxast "github.com/skycoin/cx/cx/ast"
 	"github.com/skycoin/cx/cx/astapi"
 	cxconstants "github.com/skycoin/cx/cx/constants"
-	cxgo "github.com/skycoin/cx/cxparsergenerator/cxparsingcompletor"
+	cxgo "github.com/skycoin/cx/cxparser/cxparsingcompletor"
 )
 
 func TestASTAPI_Expressions(t *testing.T) {
