+++ cxparser/actions/counter.go
