+++ cxparser/cxparsingcompletor/lexer.y
@@ -13,11 +13,11 @@ This is a machine generated file
 lexer.go is is generated by GOYACC from lexer.y
 
 build-parser: ## Generate lexer and parser for CX grammar
-  	./bin/goyacc -o cxparsergenerator/cxpartialparsing/parser.go cxparsergenerator/cxpartialparsing/grammer.y
-	./bin/goyacc -o cxparsergenerator/cxparsingcompletor/lexer.go cxparsergenerator/cxparsingcompletor/lexer.y
+  	./bin/goyacc -o cxparser/cxpartialparsing/parser.go cxparser/cxpartialparsing/grammer.y
+	./bin/goyacc -o cxparser/cxparsingcompletor/lexer.go cxparser/cxparsingcompletor/lexer.y
 
 - lexer.go is output
-- cxparsergenerator/cxparsingcompletor/lexer.y is input
+- cxparser/cxparsingcompletor/lexer.y is input
 */
         
 	// var actions.AST = MakeProgram(CALLSTACK_SIZE, STACK_SIZE, INIT_HEAP_SIZE)
